Update finalcode.R
So the code as it was before, took the first row and column from the feols output, however the function cuts out rows where that variable's coefficients have collinearity problems (or missing subsets that made the coefficient = NA); what the code was then doing was using a different variable's coefficient in place of the variable of interest's. 

The new code makes sure that the coefficient name = the variable of interest's name, or sets the beta and SD to zero otherwise.